MapSerializer: allow easier transition from non-versioned to versioned
Before when you wanted to change a non-versioned key or value to a versioned
one, the file header size would change (hash variable would disappear and
version variable would appear) and one would first have to try loading
the non-versioned version and if that fails the versioned version. And even that
is not always guaranteed to work.

This change makes it that the file header always has the same size and always
offers the  same variables. When no version info is used a default version of 0
is taken. When a struct hash can't be applied (due to a class being the
value/key), the mangled name of the type is used.

This also means that we now always do a struct hash check too, even if the
structs are versioned, that way we catch any potential human errors that might
happen while setting up the version information in the structs.