Fix erroneous behaviour of PCRE findFirst()
89600396a450311f982642e3bdd3ff74a2626023 altered some of the internal logic
of the PCRE classes and introduced a bug meaning that findFirst() would
throw an exception if multiple substrings would match. This is not an
error -- it still indicates that a match has been found and allows the
first matched substring to be returned.

This commit also clarifies the comments inside findFirst().

Fixes #710.