Prioritycache: Call drop-item notifier after dropping the item
Before this commit the event was called before dropping the item. The
problem is that a notified component X might try to take an action (e.g
populate the item again into the cache) using another component Y, however
component Y would check the cache first to see if the item is there.
Since the item is not yet deleted so the cache reports back that it still
exists although it's few code instructions away from being deleted.

Getting around this problem without fixing the cache requires the using
of schedulers to take an action after the item is removed. Fixing this
in the cache itself is much simpler and makes the class more usable.

The refactored notifier takes a reference to the dropped value as well
as the key as parameters.