D2: Correct override in io.device.MemoryDevice
Change `error` and `write` to take an `istring`, as in tango.
`toString` wasn't conforming to the interface: tango documentation
mention that it "Return the name of this conduit." whereas the
implementation returned an internal buffer cast to `char[]`.
Hence, toString has been deprecated, in favor of the more common
`peek` method.