Add isTessellatedBy method to ocean.math.Range
Predicate that checks whether the provided array of ranges exactly
tessellates this range.  The term "tessellation" means that this
range is a union of the given ranges and that the given ranges form
a contiguous chain without gap or overlap.

This function is added as main replacement the existing check
of consistency an array of ranges.

Part of sociomantic/ocen#747