Add isSupersetOf method to ocean.math.Range
Determines whether this instance is a superset of the non empty
specified range. All values in the other range must be within this range.

This function is added as replacement of the existing `supersetOf` function
as discussed in sociomantic/ocean#747. Legacy `supersetOf` was ill defined
from mathematical and practical points of view. Also new function provides
better name convention.

Part of sociomantic/ocean#747