Add isContiguous function to ocean.math.Range
Predicate that checks contiguity of the array of Range!T.

This function's result is equivalent to !hasGap && !hasOverlap.
It has been implemented as a separate function because a more efficient
implementation is possible.

This function is added as part of split the existing check
of consistency an array of ranges to separate functions.

Part of sociomantic/ocean#747