Add extent function to ocean.math.Range
Generate a single Range!T that covers the entire set of values found
in an array of Range!T, i.e. whose min, max values reflect the smallest
and largest min and max found in the array.

This function is added as part of split the existing check
of consistency an array of ranges to separate functions.

Part of sociomantic/ocean#747