D2: don't mark IAggregatePool scope classes as scope
This is a dmd2 bug workaround which caused it to refuse to compiled
nested scope definitions if those had constructors (and some other
obscure conditions).

Trying to debug and fix the issue has proved to be unreasonably effort-consuming.

Those classes still must be used as scope classes but only swarm is affected
so documented annotations should be good enough.