Remove mutation of read-only data
This (deprecated) unittest was mutating data in place. When fed literals, it could segfault (it did in D2). As the code is arguably wrong both in D1 and D2, and the functionality is deprecated and so not likely to have any modification in the future, we just remove the unittest.