Fix incorrect buffer expansion in tango.io.device.File
File.get took a second parameter of type `void[]` which would default to null.
When the length was less than the required one, it would then allocate this array.
This could lead some to believe that the buffer will be expanded, while in reality
the reallocation wouldn't reflect on the provided buffer as it wasn't passed by ref.
This commit split this function into two, to provided the expected behaviour.