Fix Range.opCall to handle creating empty ranges
An empty range (as defined by the isEmpty function) is also a valid range
(as defined by the isValid function), so should not be rejected by the in
contract of opCall.

The (fixed) check is moved from an in contract to an enforce in the
function body as it should be checked in all builds -- creating an invalid
range should never be allowed.