D2: Remove final qualifier on variable
Those variable where declared as `static final`, which is illegal in D2.
They could have been make `const` (not `static const`), but we know
that it triggers memory allocation with array literals, and since their usage
is very limited and only in this package, we just remove the `final` and leave
them as `static`.