D2: Improve serializer const correctness
In light of porting swarm, some const inconsistency where identified
in the serializer code.
First, we accept `const` data as an input for the Version decorator.
Second, the LoadCopy method from the serializer has a constant
input source (since it does copy it, as the name imply), so we reflect
that on the parameter.