Remove version dashdash
With version dashdash enabled, parameters following '--' on the
command-line would be assigned to the default (null) argument. Without
version dashdash, they would be assigned to the last known argument
target.

But version dashdash is always enabled, and no foreseeable usecase for
disabling it presents itself. Hence it is simply removed.