Fix array out of bounds error
An array out of bounds error is triggered when a long string is passed
as the first parameter to the 'timeToUnixTime()' function. This is
because of an incorrect initialization of the formatting buffer.
Reducing the length of the formatting buffer by one also means that one
less character (the separator) is matched in the call to 'sscanf()'. As
a result, the corresponding case statement within the switch block also
needs to be fixed. A unit test has also been added to cover this
testcase.

Fixes #971.