Fix the implementation of pop() in the NotifyingQueue
Multiple calls of pop() would overwrite the previous data
as an internal ContiguousBuffer was being used.

Instead, the new pop() overload now expects the user to provide
a pair of ContigousBuffer and a ubyte[] buffer when
NotifyingQueue is instantiated with a struct type.

The old version of pop() for structs has been deprecated