Deprecate tango.util.log.Trace and tango.util.log.Config
We deprecated the `Trace` module because it relies on the static constructor
order, which is not guaranteed and can change by adding or removing import,
as well as the compilation mode (release or debug).
User that wish to retain the behaviour should copy the code in their main file.

The `Config` module was just publicly importing `Trace` and providing an alias
for `Log`. Since it didn't bring any value, and is unlikely to be used, we
kill it as well.