Add run-time warning upon instantiating tango.util.log.AppendSyslog
AppendSyslog is due to be deprecated and removed (see #938). Due to some
uncertainty as to how exactly the class will be deprecated (it may turn out
that it's simply removed in a future release), a warning is added now so
that users are aware of the need to stop using it.