Deprecate IStatsLog
There is no apparent reason for the separation of StatsLog vs ISatsLog.
(The documentation refers to IStatsLog as "template-free": it seems that
StatsLog was templated at some point in the past.)

All functionality of IStatsLog is rolled into Statslog and the former
deprecated. Code in ocean which refers to IStatsLog is adapted to use
StatsLog instead.

Fixes #740.