Add StatsExt.newStatsLog()
A new method, `newStatsLog()` has been added. This method allows the user
to create `StatsLog` instances in addition to the one which is created
automatically by the `StatsExt` and stored in its `stats_log` member. The
new stats logger is configured according to the default settings and will
be registered with the `ReopenableFilesExt`, if present, for log rotation.
(The ability to create multiple stats loggers is desired by some
applications, so this method has been added to enable them to easily create
properly configured instances.)

Fixes #1062.