Deprecate allocating signatures of readline()
The previous behavior of `readline()` function was to either take a
pre-allocated buffer (and expand it if needed) or allocate a new buffer.

Both of these behaviors allocate memory and it leaves no way to pass a
reusable buffer. This commit changes the method signature such that it
takes a reusable buffer as a parameter.