Make MicrosecondsClock not implement IMicrosecondsClock
The class contained only a single non-static member: now_us(), a method
required by the IMicrosecondsClock interface. IMicrosecondsClock is going
to be deprecated, so this class must be adapted. The one non-static member
is made static; an instance is no longer required.

The one location in ocean where this class is used has been adapted.