Add ocean.net.collectd.Identifier
This struct is a wrapper for Collectd's identifier.
The identifier Collectd uses is composed to 3 to 5 parts, which are:
 - hostname
 - plugin
 - plugin instance (optional)
 - type
 - type instance (optional)

It is sent to / received from Collectd as a string with the follow pattern:
`hostname/plugin-plugin_instance/type-type_instance`
There are various rules about which characters can or cannot be in the various
part of the identifier.
In order to abstract all of this from the user, and also make the code simpler,
Collectd binding will only accept `Identifier`.

Part of sociomantic/ocean#622