D2: Simplify code in util.container.map.model.Bucket
Using `offsetof` of one of the struct's member for the definition
of another member relies on the evaluation order of the frontend,
and thus might break at some point (in fact it did in 2.071).
While this is technically a regression (issue 15965), the code
does not need that amount of complication as `offsetof` needs
to be 0 anyway.
Also, a `typeof(this)` was turned into the actual name, to
increase readability and because it confused DMD's diagnostic.