Implement ocean.core.Buffer
Wrapper struct for an array of type T which is meant to
own the data and enable stomping each time length changes. To
ensure no accidental reallocation, it is declared as non-copyable
in D2 and can only be passed by ref.

Internally implementation is split into 3 mixed in modules for
conceptually different cases - trying to keep all special
cases in one has resulted in unreadable `static if` mess.