Move sections below the module description
In DDoc the module description is an implicit section that starts after the
module summary (the first paragraph in the DDoc comment) and continues
until the first explicit section. All the contents after this section then
belongs to the section and not to the description. So new sections must be
moved to the end of the documentation.