Remove GPL license from ebtree bindings
According to the most common regulations, API/ABI are not subject to
copyright or their use is considered fair use so we can license our code
with the license we want. On top of that, the author of libebtree is
changing the license to MIT[1]. Also a small clarification to the
license section is added so users are warned that other licensing
restrictions might apply when linking to an external library.

A large text describing implementation details of ebtree is also removed
since it was a copy from what the C header has, and that might be
subject to copyright. In any case is always better to consult the
original C documentation when using binding to C libraries.

Related to #1202.

[1] https://github.com/sociomantic/ebtree/issues/7#issuecomment-224563520