Remove GPL license from readline bindings
According to the most common regulations, API/ABI are not subject to
copyright or their use is considered fair use so we can license our code
with the license we want. Also a small clarification to the license
section is added so users are warned that other licensing restrictions
might apply when linking to an external library.

Related to #1202.