Remove old license from DigitalMars
Instead we move the DigitalMars to the copyright section, in case it
wasn't there already. All the code affected here was actually written
either by Walter Bright, Don Clugston and/or Sean Kelly, all of which
already changed the license to Boost (or Tango) before, so I think is
safe to assume this code can be released with Tango license without the
special mentions to the original license pre-Tango/D2.