Change copyright to Leandro Lucarella
This module was originally written by me and put into public domain. To
make things simpler, I'm changing the copyright to me and making it Boost +
BSD licensed to match the terms in Sociomantic code (is also already part
of druntime, so is not very relevant, it should be removed after moving to
D2).