Implement initial set of Task extensions
`ocean.task.extensions` is a package with various optional extensions that
affect basic functionality and/or semantics of task class. To use them,
simply use template `TaskWith` class as a base for application task classes
and supply desired extensions as template arguments.

`SelectFiberSupport` provides a way to interact way with legacy API
expecting SelectFiber instance.

`ExceptionForwarding` makes possible to rethrow exception that was not
handled by task from context it yielded to.