Deprecate text.Regex and io.stream.Patterns
The Regex module is using some code from Phobos, and we don't want more
different licenses. On top of that we are not using that module, is
basically unmaintained and we have a replacement which is probably more
efficient, flexible and maintained: PCRE.

The Patterns modules is also unused and I don't think anyone even knows
what is it for :)

Will fix #1215 when merged to v2.x.x and we'll actually remove it.