Move explanation about release notes out of the README
Move the explanation about how to read the release notes from the
README.rst to the RELEASE_NOTES.md file, it's really only useful when
upgrading to a new version.

This also remove some obvious/redundant stuff about how to migrate and
restructure the release notes to be more clear about which version is
being released, and what was the previous version.