StringStructSerializer: Add support for ISO output
Introduced an overload of the `StringStructSerializer` serializer
which takes an array of known timestamp field names.
If a field matches one of the names and implicitly converts to `ulong`,
an ISO formatted string will be emitted in parentheses next to the value of
the field (which is assumed to be a unix timestamp).

Part of: https://github.com/sociomantic/krilltools/issues/153