Fix ebtree nodes not properly reset
The ebtree's node free list is not reset when the array pool is cleared.
When using the cache again after clearing, it ends up using the same
ebtree nodes twice, first from the erroneously not cleared free list, then
again from the pool.

The bug caused the ebtree to crash or to run into an infinite loop when the
ebtree is cleared and used again afterwards.