MessageDigest/HMAC: Replace hash(void[][]...) with calculate(ubyte[][]...)
This is to prevent undesired implicit casts of multi-dimensional arrays to
void[], which cause a wrong resulting hash/HMAC, and the error is hard to
find. An illustration of the error that happened:

  void f (void[][] items ...);
  void[][] items;
  // populate items
  f(items);

Here f() receives only one void[] argument, not items.length arguments.
This is because the call is equivalent to the following:

  void[] arg = items; // Implicit cast from void[][] to void[]
  f(arg);