Add floating point formatting code to the Float module
This brings this module closer to it's Integer counterpart,
which has a `formatter` overload capable of formatting based
on an format string.

The code is almost an identical copy of Tango's `Layout.floater` method,
which was not updated to forward to this new function, as it calls
the `integer` method for `Layout`.
The notable exception is the absence of support for 'x' and 'X' formatting.
The tango code didn't provide correct result, and it's such a corner case
that support was stripped.