Replace EpollProcess.ProcessMonitor with a singleton
The ProcessMonitor does not need a copy of the epoll instance. A consequence of
this is that we do not need to worry about it being created in unit tests with
an incorrect epoll instance. Previously it was public only as a workaround for
this issue. As it is no longer required, it can be deprecated.