Fix #1293: EpollProcess is broken for multithreaded apps
Use a signal handler instead of signalfd.

Historically, EpollProcess has used signalfd to detect when child processes
terminate. Unfortunately, signalfd does not behave in the way that you would
hope. Signals are only delivered to signalfd if they are blocked. The fatal
flaw with signalfd is that Linux signal handlers are process-wide, but the
signal masks are per-thread. If the signal is unblocked on _any_ thread,
the signal handler will be called instead.

This means that for signalfd to work, all threads in the app must block the
signal we're interested in. This applies even to threads created in external
C libraries. Of course those libraries have no idea if we are interested
in SIGCHLD or not. If we want signalfd to work, we need to make sure
all of those threads block SIGCHLD. In general this is impossible to ensure.
Thus, using a signal handler is the only reliable method to detect
termination of child processes. The signal handler is kept as simple as
possible, and does nothing except trigger a SelectEvent.