Make `theScheduler.shutdown()` more fatal
Now once scheduler enters shutdown mode, any attempt to
interact with it (i.e. scheduler new task or suspend to process events)
will kill the caller using `Task.kill` method.

This fixes the issue when new tasks could be scheduled in between
cleaning the queue and getting to the end of `eventLoop` method.