Update Makd to v1.6.0 to get `make doc`
The new Makd version uses harbored-mod to build API documentation when
the target `doc` is used.

We now need to define the test runner module to use explicitly because
the new Makd auto-detects if Ocean's test runner should be used based on
submodules and we -or course- don't have Ocean as a submodule.

* submodules/makd v1.4.0(225a078)...v1.6.0(949d0ad) (12 commits)
  > Update RELEASE_NOTES with missing changes
  > Add `doc` target to build D documentation
  > Make TEST_RUNNER_STRING overrideable
  > Clean and update RELEASE_NOTES
  > Use oceans or druntimes test runner by default
  (...)