Add asserts to verify the indent length
Add asserts at the beginning of all functions that edit the final string
struct serializer output. The asserts are used for a simple verification
of the current indentation length.

The indentation is expected to be zero in the `open()` function as the
output hasn't been modified yet. In all other functions that edit the
serializer output, the indentation is expected to be non-zero.