Allow equivalent strings for whitespace characters
Single character fields in structs are currently simply serialized into
their character equivalents. However, this does not always work well if
the serialization is being done in order to print the serialized struct
on a terminal. This is because whitespace characters become invisible
and other unprintable characters generate some random output.

To overcome this, an additional boolean flag has been introduced in the
'serialize()' call. When this flag is set, whitespace or unprintable
characters will be serialized into their equivalent friendly string
representations.
For example, the string '\n' will be generated for the newline
character, '\t' for the tab character and so on.