Arguments: Avoid omitting argument without help
When setting up a command-line argument, if its help text is not given,
then the entire argument gets omitted from the help message of the
application. To avoid this, remove the early return from the
'displayArgumentHelp()' function, and avoid calling this function
altogether for the null argument (which never has an associated help
text).

Fixes #1469.