Arguments: Restructure 'space()' function
If the very first call to the 'space()' function passes zero as the
'width' parameter, then this function will return a null array.

With the new ocean Formatter, null arrays are formatted as the string
"null" instead of an empty string. This could potentially cause the
first line under the "Program options:" section to be misformatted.

To avoid this, the 'space()' function is slightly restructured so that
it can never return a null array and its return type is changed to
`cstring` (its functionality remains unaffected).