Logger: Move the Context interface in ILogger
In order to introduce a new logger class, we first need to untangle the Log module.
One of the goal is to be able to freely mix old and new logger, so things like configuration will be transparent.
This is the first step in such a direction, moving this interface to a higher level so that we can separate the Hierarchy class, which inherits from it.