Add the ability to supply a custom format string
Add a new parameter to the 'formatTime()' function that can be used to
specify a custom format string to use when formatting the given UNIX
timestamp.

When allowing the caller to set the format string, it becomes the
responsibility of the caller to ensure that the destination output
buffer is large enough to accommodate the result of the conversion. This
makes the in-contract check for whether the destination buffer is large
enough meaningless.