Deprecate OutputDevice in favour of IODevice
OutputDevice isn't used anywhere but a class that is both IInputDevice and
IOutputDevice will be needed. IODevice is InputDevice and IOutputDevice.
The reason for using InputDevice and not OutputDevice is that sockets don't
actually use write(), instead they use send() to pass a flag to suppress
SIGPIPE. They do use read() though.