ocean.util.log.Config: Simplify configureLoggers code a bit
Having booleans that controls the behavior of a function is usually a red flag.

In this case, one of the boolean enabled loose parsing,
which is a global setting and as such can be trivially taken out of the function/

The other decided on the kind of Appender to use, between two options,
which can be replaced by a delegate instantiating the Appender,
which is both cleaner and more abstract.