ConfigExt: Do not always add the defaults config files
ConfigExt would always add the defaults argument to the list of arguments,
so using `./daemonapp -c etc/other-config.ini` would result in
`[ "etc/config.ini", "etc/other-config.ini" ]` instead of `[ "etc/other-config.ini" ]`.
This problem was invisible to user code because all of them use a single configuration file,
and a workaround was put in place in `DaemonApp` which hid the bug.