Document random number requirements for `shuffle`
This change improves the documentation of the `shuffle` function
accepting a random number generator. It updates the documentation to
include that the generated numbers must be in the open interval (-1.0,
1.0).

This important to a user of this function because the information is
necessary to ensure correct functioning of the shuffling. Otherwise out
of bound accesses may be performed.

Part of #1611